refactor: consolidate local module splits into tracked branch
Split monolithic files into focused submodules:
- workflows.rs → workflows/ (listing, execution, history, models, tests)
- db.rs → db/ (connection, migrations, pragmas, tests)
- cmd/team.rs → cmd/team/ (per-subcommand modules)
- recipe_bridge/tests.rs → recipe_bridge/tests/ (focused test suites)
- Add PR Merge Manager agent instructions
