+++ crates/opengoose-cli/src/cmd/team/run.rs
@@ -1,0 +1,22 @@
+use std::sync::Arc;
+
+use anyhow::Result;
+
+use opengoose_persistence::Database;
+use opengoose_types::EventBus;
+
+pub(super) async fn run(team_name: &str, input: &str, model: Option<String>) -> Result<()> {
+    let db = Arc::new(Database::open()?);
+    let event_bus = EventBus::new(256);
+
+    println!("Running team '{team_name}'...");
+
+    let (run_id, result) =
+        opengoose_teams::run_headless_with_model(team_name, input, db, event_bus, model).await?;
+
+    println!("\n--- Result ---");
+    println!("{result}");
+    println!("\nRun ID: {run_id}");
+
+    Ok(())
+}
